move ParticleContainer into it's own header file
