+++ Src/Particle/Make.package
@@ -1,6 +1,7 @@
 
 CEXE_headers += AMReX_Particle.H
 CEXE_headers += AMReX_Particles.H
+CEXE_headers += AMReX_ParticleContainer.H
 CEXE_headers += AMReX_ParGDB.H
 CEXE_headers += AMReX_ParticleContainerI.H
 CEXE_headers += AMReX_ParticleInit.H
